Remove unneeded indirection
