+++ cypress/plugins/nodeManager/commands.ts
@@ -1,44 +1,23 @@
-import type {
-  ConnectNodesOptions,
-  OnboardNodeOptions,
-  OnboardedNode,
-  StartNodeOptions,
-} from "./shared";
+import type { OnboardedNode } from "./shared";
 
 import { Commands } from "./shared";
-
-const startNode = (options: StartNodeOptions): void => {
-  cy.task(Commands.StartNode, options);
-};
-
-const onboardNode = (options: OnboardNodeOptions): void => {
-  cy.task(Commands.OnboardNode, options);
-};
-
-const connectNodes = (options: ConnectNodesOptions): void => {
-  cy.task(Commands.ConnectNodes, options);
-};
-
-const stopAllNodes = (): void => {
-  cy.task(Commands.StopAllNodes);
-};
 
 export const withTwoConnectedNodes = (
   callback: (node0: OnboardedNode, node1: OnboardedNode) => void
 ): void => {
-  startNode({ id: 17000 });
-  onboardNode({ id: 17000 });
+  cy.task(Commands.StartNode, { id: 17000 });
+  cy.task(Commands.OnboardNode, { id: 17000 });
 
-  startNode({ id: 18000 });
-  onboardNode({ id: 18000 });
+  cy.task(Commands.StartNode, { id: 18000 });
+  cy.task(Commands.OnboardNode, { id: 18000 });
 
-  connectNodes({ nodeIds: [17000, 18000] });
+  cy.task(Commands.ConnectNodes, { nodeIds: [17000, 18000] });
 
   cy.task<OnboardedNode[]>(Commands.GetOnboardedNodes).then(nodes => {
     callback(nodes[0], nodes[1]);
   });
 
-  stopAllNodes();
+  cy.task(Commands.StopAllNodes);
 };
 
 export const asNode = (node: OnboardedNode): void => {
