+++ cypress/plugins/nodeManager/shared.ts
@@ -1,24 +1,12 @@
-export type NodeId = number;
-
-export interface StartNodeOptions {
-  id: NodeId;
-  proxyBinaryPath?: string;
-}
-
-export interface OnboardNodeOptions {
-  id: NodeId;
-  handle?: string;
-  passphrase?: string;
-}
+// Types shared between `plugin.ts` and `commands.ts`. These types can't be in
+// `plugin.ts`, because Cypress plugins run in a separate Nodejs environment
+// and can directly use Nodejs libraries, the Cypress tests don't have access
+// to those, so indlucing `plugin.ts` inside `commands.ts` leads to errors.
 
 export interface OnboardedNode {
   authToken: string;
   peerAddress: string;
   httpPort: number;
-}
-
-export interface ConnectNodesOptions {
-  nodeIds: NodeId[];
 }
 
 export enum Commands {
